Flag Fix
Установил я ваши флаги компиляции и  все покрылось предупреждениями, но вроде пофиксил
